cambios en el codigo de acuerdo a la nueva query

+++ MiAlmacen/MiAlmacen.API/Controllers/VentaController.cs
@@ -34,7 +34,7 @@ namespace MiAlmacen.API.Controllers
 
 
         [HttpPut("{pago}")]
-        public async Task<IActionResult> PutSaldo(VentaModel venta, float pago)
+        public async Task<IActionResult> PutSaldo(VentaModel venta, decimal pago)
         {
             var nuevoSaldo = _repository.PutSaldo(venta, pago);
             return Ok(nuevoSaldo);
