+++ MiAlmacen/MiAlmacen.Data/Entities/Compras.cs
@@ -13,8 +13,8 @@ namespace MiAlmacen.Data.Entities
         public int Proveedor_Id { get; set; }
         public int NroRecibo { get; set; }
         public int Empleado_Id { get; set; }
-        public float Total { get; set; }
-        public float Saldo { get; set; }
+        public decimal Total { get; set; }
+        public decimal Saldo { get; set; }
         public DateTime Fecha_Baja { get; set; }
 
         public Proveedores Proveedor { get; set; }
