+++ MiAlmacen/MiAlmacen.Data/Entities/DetalleCompras.cs
@@ -11,8 +11,8 @@ namespace MiAlmacen.Data.Entities
         public int Id { get; set; }
         public int Articulo_Id { get; set; }
         public int Cantidad { get; set; }
-        public float Precio_Mayor { get; set; }
-        public float Precio_Unit { get; set; }
+        public decimal Precio_Mayor { get; set; }
+        public decimal Precio_Unit { get; set; }
         public int Compra_Id { get; set; }
 
         public Articulos Articulo { get; set; }
