+++ MiAlmacen/MiAlmacen.Data/Entities/DetalleVentas.cs
@@ -10,10 +10,10 @@ namespace MiAlmacen.Data.Entities
     {
         public int Id { get; set; }
         public int Articulo_Id { get; set; }
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
         public int Cantidad { get; set; }
         public int Venta_Id { get; set; }
-        public float SubTotal { get; set; }
+        public decimal SubTotal { get; set; }
 
         public Articulos Articulo { get; set; } = new();  
     }
