+++ MiAlmacen/MiAlmacen.Model/Models/ArticuloModel.cs
@@ -17,9 +17,9 @@ namespace MiAlmacen.Model.Models
         [Range(0, 99999999999999, ErrorMessage = "Solo números de hasta 13 dígitos")]
         public long Codigo_Art { get; set; }
         [Required]
-        public float Precio_Unit { get; set; }
+        public decimal Precio_Unit { get; set; }
         [Required]
-        public float Precio_Mayor { get; set; }
+        public decimal Precio_Mayor { get; set; }
         [Required]
         [Range(-30, 1000, ErrorMessage = "Solo números de entre -30 y 1000")]
         public int Stock_Act { get; set; }
