+++ MiAlmacen/MiAlmacen.Model/Models/VentaModel.cs
@@ -12,9 +12,9 @@ namespace MiAlmacen.Model.Models
         public int Cliente_Id { get; set; }
         public int Empleado_Id { get; set; }
         [Required]
-        public float Total { get; set; }
+        public decimal Total { get; set; }
         [Required]
-        public float Saldo { get; set; }
+        public decimal Saldo { get; set; }
         public DateTime? Fecha_Baja { get; set; }
 
         public List<DetalleVentaModel> Detalle { get; set; } = new ();
